# feature-flag: clean up segment references on promise resolve