--- a/lib/feature_flags.js
+++ b/lib/feature_flags.js
@@ -15,7 +15,8 @@ exports.prerelease = {
   fastify_instrumentation: false,
   // Starts by defaulting true as introducing for deprecation/removal purposes.
   // Will eventually set false prior full removal of feature.
-  certificate_bundle: true
+  certificate_bundle: true,
+  new_promise_tracking: false
 }
 
 // flags that are no longer used for released features