--- a/test/integration/core/async_hooks-new-promise.tap.js
+++ b/test/integration/core/async_hooks-new-promise.tap.js
@@ -6,7 +6,10 @@
 'use strict'
 
 const exec = require('child_process').execSync
-exec('node --expose-gc ./async_hooks.js', {
-  stdio: 'inherit',
-  cwd: __dirname
-})
+exec(
+  'NEW_RELIC_FEATURE_FLAG_NEW_PROMISE_TRACKING=1 node --expose-gc ./async_hooks-new-promise.js',
+  {
+    stdio: 'inherit',
+    cwd: __dirname
+  }
+)